[#522] - On creation and deletion of meal plan refetch the tags to show the updated tags
