+++ mealplanner-ui/package.json
@@ -78,5 +78,5 @@
     "@types/react-relay": "^13.0.1",
     "@types/relay-runtime": "^13.0.1"
   },
-  "proxy": "http://localhost:4000"
+  "proxy": "http://127.0.0.1:4000"
 }
