+++ mealplanner-ui/src/pages/MealPlans/MealPlans.tsx
@@ -1,57 +1,20 @@
-import { ContentCopy, DeleteTwoTone, ShoppingCart } from "@mui/icons-material";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import FavoriteIcon from "@mui/icons-material/Favorite";
 import {
-  Avatar,
-  Card,
-  CardActions,
-  CardContent,
-  CardHeader,
-  Collapse,
   FormControl,
   FormControlLabel,
   Grid,
-  IconButton,
-  IconButtonProps,
-  ImageList,
-  ImageListItem,
   InputBase,
   Radio,
-  RadioGroup,
-  Typography,
-  styled
+  RadioGroup
 } from "@mui/material";
 import { graphql } from "babel-plugin-relay/macro";
-import React, { useState } from "react";
-import { useLazyLoadQuery } from "react-relay";
-import { useNavigate } from "react-router";
-import { getCurrentPerson } from "../../state/state";
-import { MealPlanNode } from "../../state/types";
+import { Suspense, useState } from "react";
+import { useLazyLoadQuery, useRefetchableFragment } from "react-relay";
 import { CreateMealPlan } from "./CreateMealPlan";
-import { deleteMealPlan } from "./DeleteMealPlan";
-import { duplicateMealPlan } from "./DuplicateMealPlan";
-import { MealPlansTags } from "./MealPlansTags";
+import { MealPlanCard } from "./MealPlanCard";
+import { MealPlansTags, MealPlansTagsFragment } from "./MealPlansTags";
 import { MealPlansQuery } from "./__generated__/MealPlansQuery.graphql";
 
-interface ExpandMoreProps extends IconButtonProps {
-  expand: boolean;
-}
 
-const ExpandMore = styled((props: ExpandMoreProps) => {
-  const { expand, ...other } = props;
-  return <IconButton {...other} />;
-})(({ theme, expand }) => ({
-  transform: !expand ? "rotate(0deg)" : "rotate(180deg)",
-  marginLeft: "auto",
-  transition: theme.transitions.create("transform", {
-    duration: theme.transitions.duration.shortest,
-  }),
-}));
-
-interface MealPlanCardProps {
-  mealplan: MealPlanNode;
-  connection: string;
-}
 
 const mealPlansQuery = graphql`
   query MealPlansQuery {
@@ -88,137 +51,6 @@ const mealPlansQuery = graphql`
   }
 `;
 
-
-
-const getInitials = (name: string) => {
-  let initials = "";
-  let names: string[] = (name && name.length > 1 && name.split(" ")) || [
-    "No",
-    "Name",
-  ];
-  names.forEach((n) => {
-    initials += n[0];
-  });
-  return initials;
-};
-
-const MealPlanCard = (props: MealPlanCardProps) => {
-  const [expanded, setExpanded] = React.useState(false);
-  const navigate = useNavigate();
-  const mealplan = props.mealplan;
-  const connection = props.connection;
-  const handleExpandClick = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    setExpanded(!expanded);
-  };
-
-  return (
-    <Grid item xs="auto">
-      <Card
-        sx={{ maxWidth: 332 }}
-        onClick={(e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          navigate(`/mealplans/${mealplan.rowId}`);
-        }}
-        onMouseOver={(e) => {
-          e.currentTarget.style.cursor = "pointer";
-        }}
-      >
-        <CardHeader
-          avatar={
-            <Avatar sx={{ bgcolor: "green", width: "fit" }} aria-label="user">
-              {getInitials(mealplan.person?.fullName || "")}
-            </Avatar>
-          }
-          action={
-            <div>
-              <IconButton
-                aria-label="shopping list"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  console.log("stopped propagation");
-                  navigate(`/mealplans/${mealplan.rowId}/shopping-list`);
-                }}
-              >
-                <ShoppingCart />
-              </IconButton>
-              <IconButton
-                aria-label="delete"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  console.log("meal plan id: ", typeof mealplan.rowId);
-                  deleteMealPlan(connection, mealplan.rowId);
-                }}
-              >
-                <DeleteTwoTone />
-              </IconButton>
-              {getCurrentPerson().personRole === "app_admin" || getCurrentPerson().personRole === "app_meal_designer" ? (
-                <IconButton
-                  aria-label="duplicate"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    duplicateMealPlan(connection, mealplan.rowId);
-                  }}
-                >
-                  <ContentCopy />
-                </IconButton>
-              ):null}
-            </div>
-          }
-          title={mealplan.nameEn}
-          subheader={mealplan.person?.fullName}
-        />
-        <ImageList sx={{ width: 350, height: 150 }} cols={3} rowHeight={164}>
-          {mealplan.mealPlanEntries.nodes.map((meal) =>
-            meal.meal?.photoUrl !== null ? (
-              <ImageListItem key={meal.meal?.id}>
-                <img
-                  src={`${meal.meal?.photoUrl}`}
-                  srcSet={`${meal.meal?.photoUrl}`}
-                  alt={meal.meal?.photoUrl || "no image"}
-                  loading="lazy"
-                />
-              </ImageListItem>
-            ) : (
-              <></>
-            )
-          )}
-        </ImageList>
-        <CardContent>
-          <Typography variant="body2" color="text.secondary">
-            {mealplan.tags?.map((tag) => (
-              <span>{tag} &nbsp;</span>
-            ))}
-          </Typography>
-        </CardContent>
-        <CardActions disableSpacing>
-        <IconButton aria-label="add to favorites">
-            <FavoriteIcon />
-          </IconButton>
-
-          <ExpandMore
-            expand={expanded}
-            onClick={handleExpandClick}
-            aria-expanded={expanded}
-            aria-label="show more"
-          >
-            <ExpandMoreIcon />
-          </ExpandMore>
-        </CardActions>
-        <Collapse in={expanded} timeout="auto" unmountOnExit>
-          <CardContent>
-            <Typography>
-              {" "}
-              <div>{mealplan.descriptionEn}</div>
-            </Typography>
-          </CardContent>
-        </Collapse>
-      </Card>
-    </Grid>
-  );
-};
-
 export const MealPlans = () => {
   const [searched, setSearched] = useState<string>("");
   const [searchType, setSearchType] = useState('name');
@@ -228,6 +60,9 @@ export const MealPlans = () => {
     {},
     { fetchPolicy: "network-only" }
   );
+
+  const [_, refetch] = useRefetchableFragment(MealPlansTagsFragment, data);
+  
   const selectedTags = data.gqLocalState.selectedMealPlanTags || [];
 
   return (
@@ -290,14 +125,16 @@ export const MealPlans = () => {
         </FormControl>
         <span>
         {data.mealPlans ? (
-          <CreateMealPlan connection={data.mealPlans?.__id} />
+          <CreateMealPlan connection={data.mealPlans?.__id} refetch={refetch}  />
         ) : (
           <></>
         )}
         </span>
         </Grid>
         {searchType === 'tags' &&
+          <Suspense fallback={"loading tags..."}>
             <MealPlansTags tags={data}/>
+          </Suspense>
         }
       {data.mealPlans ? (
         <Grid container spacing={2} margin="1rem" columns={4}>
@@ -306,6 +143,7 @@ export const MealPlans = () => {
               return (
                 <MealPlanCard
                   mealplan={node}
+                  refetch={refetch}
                   connection={data.mealPlans!.__id}
                 />
               );
